@@ -1,15 +1,14 @@
-﻿namespace SampleMapperApp.ConsoleApp.DataModels
+﻿namespace SampleMapperApp.ConsoleApp.DataModels;
+
+public class AddressModel
 {
-    public class AddressModel
-    {
-        public int PersonId { get; set; }
+    public int PersonId { get; set; }
         
-        public string StreetNumber { get; set; }
+    public string StreetNumber { get; set; }
         
-        public string StreetName { get; set; }
+    public string StreetName { get; set; }
         
-        public string ZipCode { get; set; }
+    public string ZipCode { get; set; }
         
-        public string Country { get; set; }
-    }
+    public string Country { get; set; }
 }
