@@ -2,36 +2,35 @@
 using SampleMapperApp.ConsoleApp.DataModels;
 using SampleMapperApp.ConsoleApp.Domain;
 
-namespace SampleMapperApp.ConsoleApp.Maps
+namespace SampleMapperApp.ConsoleApp.Maps;
+
+public class PersonMap : IMap<Person, PersonModel>, IMap<PersonModel, Person>
 {
-    public class PersonMap : IMap<Person, PersonModel>, IMap<PersonModel, Person>
+    private readonly IMapper _mapper;
+
+    public PersonMap(IMapper mapper)
     {
-        private readonly IMapper _mapper;
-
-        public PersonMap(IMapper mapper)
-        {
-            _mapper = mapper;
-        }
+        _mapper = mapper;
+    }
         
-        public PersonModel Map(Person source)
+    public PersonModel Map(Person source)
+    {
+        return new PersonModel
         {
-            return new PersonModel
-            {
-                Id = source.Id,
-                FirstName = source.FirstName,
-                LastName = source.LastName,
+            Id = source.Id,
+            FirstName = source.FirstName,
+            LastName = source.LastName,
                 
-                // Calling the mapper to map the address to the model.
-                AddressModel = _mapper.Map<Address, AddressModel>(source.Address)
-            };
-        }
+            // Calling the mapper to map the address to the model.
+            AddressModel = _mapper.Map<(Address, int), AddressModel>((source.Address, source.Id))
+        };
+    }
 
-        public Person Map(PersonModel source)
-        {
-            // Convert the address model back to an address.
-            var address = _mapper.Map<AddressModel, Address>(source.AddressModel);
+    public Person Map(PersonModel source)
+    {
+        // Convert the address model back to an address.
+        var address = _mapper.Map<AddressModel, Address>(source.AddressModel);
             
-            return new Person(source.Id, source.FirstName, source.LastName, address);
-        }
+        return new Person(source.Id, source.FirstName, source.LastName, address);
     }
 }
